Separate IIB classes to the new files
